requirements.txt
eth-brownie==1.18.1
web3==5.27.0
pytest==6.2.5
vyper==0.3.1